Move mapper to folder
